submission/requirements.txt
albumentations==0.5.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
efficientnet-pytorch==0.7.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.4
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.2.0.32
opencv-python-headless==4.5.1.48
pandas==1.1.5
Pillow==8.2.0
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
